refactor(test-harness): one type per file — split input Report
input/mod.rs held Input and Report. Report and its seven From conversions move
to input/report.rs; Input keeps the module root, since the confirmed plan keeps
it there as the input type.
