+++ solx-benchmark-converter/src/input/mod.rs
@@ -7,21 +7,15 @@ pub mod compilation_time;
 pub mod error;
 pub mod foundry_gas;
 pub mod foundry_size;
+pub mod report;
 pub mod test_failures;
 pub mod testing_time;
 
 use std::path::Path;
 use std::path::PathBuf;
 
-use crate::benchmark::Benchmark;
-
-use self::build_failures::BuildFailuresReport;
-use self::compilation_time::CompilationTimeReport;
-use self::error::Error as InputError;
-use self::foundry_gas::FoundryGasReport;
-use self::foundry_size::FoundrySizeReport;
-use self::test_failures::TestFailuresReport;
-use self::testing_time::TestingTimeReport;
+use crate::input::error::Error as InputError;
+use crate::input::report::Report;
 
 ///
 /// Benchmark input format.
@@ -53,70 +47,6 @@ impl Input {
             project: project.into(),
             toolchain: toolchain.into(),
         }
-    }
-}
-
-///
-/// Enum representing various benchmark formats from tooling.
-///
-#[derive(Debug, serde::Deserialize)]
-#[serde(untagged)]
-pub enum Report {
-    /// Benchmark converter's native benchmark report format.
-    Native(Benchmark),
-    /// Foundry gas report.
-    FoundryGas(FoundryGasReport),
-    /// Foundry size report.
-    FoundrySize(FoundrySizeReport),
-    /// Compilation time report.
-    CompilationTime(CompilationTimeReport),
-    /// Testing time report.
-    TestingTime(TestingTimeReport),
-    /// Build failures report.
-    BuildFailures(BuildFailuresReport),
-    /// Test failures report.
-    TestFailures(TestFailuresReport),
-}
-
-impl From<Benchmark> for Report {
-    fn from(report: Benchmark) -> Self {
-        Self::Native(report)
-    }
-}
-
-impl From<FoundryGasReport> for Report {
-    fn from(report: FoundryGasReport) -> Self {
-        Self::FoundryGas(report)
-    }
-}
-
-impl From<FoundrySizeReport> for Report {
-    fn from(report: FoundrySizeReport) -> Self {
-        Self::FoundrySize(report)
-    }
-}
-
-impl From<CompilationTimeReport> for Report {
-    fn from(report: CompilationTimeReport) -> Self {
-        Self::CompilationTime(report)
-    }
-}
-
-impl From<TestingTimeReport> for Report {
-    fn from(report: TestingTimeReport) -> Self {
-        Self::TestingTime(report)
-    }
-}
-
-impl From<BuildFailuresReport> for Report {
-    fn from(report: BuildFailuresReport) -> Self {
-        Self::BuildFailures(report)
-    }
-}
-
-impl From<TestFailuresReport> for Report {
-    fn from(report: TestFailuresReport) -> Self {
-        Self::TestFailures(report)
     }
 }
